MG-478: reorganize CT components to be consistently nested by feature
