+++ libs/explorers/comparison-tool/src/lib/comparison-tool-controls/comparison-tool-search-input/comparison-tool-search-input.component.ts
