+++ libs/explorers/comparison-tool/src/lib/comparison-tool-controls/displayed-results/displayed-results.component.spec.ts
