+++ libs/explorers/comparison-tool/src/lib/comparison-tool-table/comparison-tool-columns/comparison-tool-columns.component.scss
