+++ libs/explorers/comparison-tool/src/lib/comparison-tool-table/comparison-tool-table-link/comparison-tool-table-link.component.html
