+++ libs/explorers/comparison-tool/src/lib/comparison-tool-table/comparison-tool-table.component.ts
@@ -16,8 +16,8 @@ import {
 } from '@sagebionetworks/explorers/services';
 import { SvgIconComponent } from '@sagebionetworks/explorers/util';
 import { TooltipModule } from 'primeng/tooltip';
-import { ComparisonToolColumnsComponent } from '../comparison-tool-columns/comparison-tool-columns.component';
 import { BaseTableComponent } from './base-table/base-table.component';
+import { ComparisonToolColumnsComponent } from './comparison-tool-columns/comparison-tool-columns.component';
 
 @Component({
   selector: 'explorers-comparison-tool-table',
